Moved common API methods into API file
